Skip Blob tests on node 16
