+++ langchain/src/document_loaders/tests/csv.test.ts
@@ -1,6 +1,5 @@
 import url from "node:url";
 import path from "node:path";
-import fs from "node:fs/promises";
 import { test, expect } from "@jest/globals";
 import { CSVLoader } from "../csv.js";
 import { Document } from "../../document.js";
@@ -23,32 +22,6 @@ test("Test CSV loader from file with column arg", async () => {
   expect(docs[1]).toEqual(
     new Document({
       metadata: { source: filePath, line: 2 },
-      pageContent: "<i>Reunited, Rush Clovis and Senator Amidala</i>",
-    })
-  );
-});
-
-test("Test CSV loader from blob with column arg", async () => {
-  const filePath = path.resolve(
-    path.dirname(url.fileURLToPath(import.meta.url)),
-    "./example_data/Star_Wars_The_Clone_Wars_S06E07_Crisis_at_the_Heart.csv"
-  );
-  const loader = new CSVLoader(
-    new Blob([await fs.readFile(filePath)], { type: "text/csv" }),
-    "html"
-  );
-  const docs = await loader.load();
-  expect(docs.length).toBe(32);
-  expect(docs[0]).toEqual(
-    new Document({
-      metadata: { source: "blob", blobType: "text/csv", line: 1 },
-      pageContent:
-        "<i>Corruption discovered at the core of the Banking Clan!</i>",
-    })
-  );
-  expect(docs[1]).toEqual(
-    new Document({
-      metadata: { source: "blob", blobType: "text/csv", line: 2 },
       pageContent: "<i>Reunited, Rush Clovis and Senator Amidala</i>",
     })
   );
